Making the list scrollable
Wrapping the transaction list with a SingleChildScrollView allows for the list to be scrollable itself, but needs to know the height from its parent container to do so.

The purpose for wrapping the main app body with SingleChildScrollView simply allows for the entire page itself to be scrollable, which helps resolve the overflow view when the keyboard pops up from the bottom when adding a new transaction.